phases 4-7
