+++ neighbourhood_observations_test.go
