+++ stash_tracker.go
